Improve the GUI
